+++ pkg/scalers/gcp_cloud_tasks_scaler.go
@@ -110,13 +110,12 @@ func parseGcpCloudTasksMetadata(config *scalersconfig.ScalerConfig) (*gcpCloudTa
 
 func (s *gcpCloudTasksScaler) Close(context.Context) error {
 	if s.client != nil {
-		err := s.client.MetricsClient.Close()
+		err := s.client.Close()
 		s.client = nil
 		if err != nil {
 			s.logger.Error(err, "error closing StackDriver client")
 		}
 	}
-
 	return nil
 }